Tests for EpochToRelativeTimeConverter

+++ app/src/main/java/com/caiolandau/devigetredditclient/domain/api/response/RedditPostsResponse.kt
@@ -55,7 +55,7 @@ class RedditPostsResponsePostData(
     val thumbnail: String?,
 
     @Json(name = "created_utc")
-    val createdUtc: Double,
+    val createdUtc: Long,
 
     @Json(name = "url")
     val url: String?,
